services/mask-generation/requirements.txt
# requirements.txt
numpy>=1.23.5,<2.0.0
transformers
pillow
fastapi
uvicorn
redis
easyocr
pynvml
python-multipart